refactor: rewrote plant triangle fragment
rewrote plant triangle fragment to be modular for easy extension and maintenance

BREAKING CHANGE:
