+++ app/src/main/java/com/akilimo/rya/views/fragments/ui/TriangleFragment.kt
@@ -2,15 +2,12 @@ package com.akilimo.rya.views.fragments.ui
 
 import android.content.Context
 import android.os.Bundle
-import android.text.Editable
 import android.text.InputType
-import android.text.TextWatcher
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.LinearLayout
 import androidx.appcompat.widget.AppCompatEditText
-import androidx.core.view.marginTop
 import androidx.core.view.setPadding
 import androidx.fragment.app.Fragment
 import com.akilimo.rya.AppDatabase
@@ -23,38 +20,27 @@ import com.google.android.material.snackbar.Snackbar
 import com.google.android.material.textfield.TextInputLayout
 import com.stepstone.stepper.VerificationError
 
-
-private const val PLANT_COUNT = "plant_count"
-private const val TRIANGLE_NAME = "triangle_name"
-
 /**
  * A simple [Fragment] subclass.
  * Use the [TriangleFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class TriangleFragment : BaseStepFragment() {
-    var triangleCount: Int = -1
-    var triangleName: String? = null
+class TriangleFragment : BasePlantTriangle() {
 
     private var _binding: FragmentTriangleBinding? = null
     private var ctx: Context? = null
 
-    private var database: AppDatabase? = null
     private val binding get() = _binding!!
-
-    private val inputLayouts: MutableList<TextInputLayout> = arrayListOf()
 
     companion object {
         /**
-         * @param plantCount Parameter 1.
+         * @param triangleCount Parameter 1.
          * @return A new instance of fragment TriangleFragment.
          */
         @JvmStatic
-        fun newInstance(plantCount: Int, triangleName: String) = TriangleFragment().apply {
-            arguments = Bundle().apply {
-                putInt(PLANT_COUNT, plantCount)
-                putString(TRIANGLE_NAME, triangleName)
-            }
+        fun newInstance(triangleCount: Int, triangleName: String) = TriangleFragment().apply {
+            this.triangleName = triangleName
+            this.triangleCount = triangleCount
         }
 
         @JvmStatic
@@ -68,10 +54,6 @@ class TriangleFragment : BaseStepFragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        arguments?.let {
-            triangleCount = it.getInt(PLANT_COUNT)
-            triangleName = it.getString(TRIANGLE_NAME)
-        }
         database = AppDatabase.getDatabase(ctx!!)
     }
 
@@ -82,8 +64,8 @@ class TriangleFragment : BaseStepFragment() {
         return binding.root
     }
 
-    override fun onSelected() {
-        super.onSelected()
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
         val lyt = binding.lytTextField
         lyt.removeAllViews() //clear all components
         for (i in 0 until triangleCount) {
@@ -105,88 +87,9 @@ class TriangleFragment : BaseStepFragment() {
     }
 
 
-    private fun addTextInputLayout(identifier: Int, context: Context): TextInputLayout {
-
-        val layoutParams = LinearLayout.LayoutParams(
-            /* width = */ ViewGroup.LayoutParams.MATCH_PARENT,
-            /* height = */ ViewGroup.LayoutParams.WRAP_CONTENT
-        )
-        layoutParams.setMargins(0, 0, 0, resources.getDimension(R.dimen.dimen_124).toInt())
-
-        val textInputLayout = TextInputLayout(
-            context, null, R.attr.customTextInputStyle
-        )
-        textInputLayout.id = identifier
-        textInputLayout.hint = getString(R.string.lbl_root_weight_hint)
-        textInputLayout.endIconMode = TextInputLayout.END_ICON_CLEAR_TEXT
-        //textInputLayout.layoutParams = layoutParams
-
-
-        textInputLayout.addView(addEditText(textInputLayout.context))
-
-        return textInputLayout
-    }
-
-
-    private fun addEditText(context: Context): AppCompatEditText {
-        val layoutParams = LinearLayout.LayoutParams(
-            /* width = */ ViewGroup.LayoutParams.MATCH_PARENT,
-            /* height = */ ViewGroup.LayoutParams.WRAP_CONTENT
-        )
-        layoutParams.setMargins(
-            /* left = */ 0,
-            /* top = */0,
-            /* right = */0,
-            /* bottom = */resources.getDimension(R.dimen.spacing_medium).toInt()
-        )
-
-
-        val editText = AppCompatEditText(context)
-//        editText.width = ViewGroup.LayoutParams.MATCH_PARENT
-//        editText.height = ViewGroup.LayoutParams.WRAP_CONTENT
-        editText.layoutParams = layoutParams
-        editText.setPadding(resources.getDimension(R.dimen.dimen_16).toInt())
-        editText.inputType = InputType.TYPE_NUMBER_FLAG_DECIMAL or InputType.TYPE_CLASS_NUMBER
-        return editText
-    }
-
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
-    }
-
-    override fun verifyStep(): VerificationError? {
-        var inputValid = false
-        val plantTrianglesMeasurement: MutableList<PlantTriangleEntity> = arrayListOf()
-        var plantNumber = 1
-        for (inputLayout in inputLayouts) {
-            val rootWeightString = inputLayout.editText?.editableText.toString()
-            val rootWeight = StringToNumberFactory.stringToDouble(rootWeightString)
-            inputValid = rootWeight > 0
-            if (inputValid) {
-                //save this value to the database
-                inputLayout.error = null
-                plantTrianglesMeasurement.add(
-                    PlantTriangleEntity(
-                        triangleName = triangleName!!,
-                        plantName = "plant$plantNumber",
-                        rootWeight = rootWeight
-                    )
-                )
-                plantNumber++
-            } else {
-                inputLayout.error = "Provide correct plant root weight"
-                inputLayout.requestFocus()
-                break //no need to loop all through
-            }
-        }
-
-        if (!inputValid) {
-            return VerificationError("Provide correct plant root weight for all inputs")
-        }
-
-        database?.plantTriangleDao()?.insertAll(plantTrianglesMeasurement)
-        return verificationError
     }
 
     override fun onError(error: VerificationError) {
@@ -200,7 +103,7 @@ class TriangleFragment : BaseStepFragment() {
         snackBar.show()
     }
 
-    @Deprecated("To be removed")
+    @Deprecated("To be removed", ReplaceWith("false"))
     fun validateInput(): Boolean {
         return false
     }
