+++ nextrelease.txt
@@ -1,1 +1,1 @@
-2.10.6
+2.10.11
